Backfill types for some dependencies
Some of our internal NPM packages are still untyped. Currently we get
around this by stubbing the types for these modules in a
`dependencies.d.ts` file, but this merely tags everything exported from
the modules as `any`.

To address this, this commit adds proper types for most of the modules
marked in `dependencies.d.ts` so that these types can be accessed and
utilized within the code that uses these modules. The modules that are
left over are those that would either take too long to fill in or those
that we are planning on removing soon anyway.